feat: add gracefull server and socket close
