+++ .gitignore
@@ -1,6 +1,6 @@
 # dependencies
 /node_modules
-
+/dist
 # System Files
 .DS_Store
 Thumbs.db
