+++ index.ts
@@ -1,6 +1,7 @@
-import { httpServer } from './src/http_server/index.js';
 // @ts-ignore
-import { wsInit } from './src/ws/websocket.ts';
+import { httpServer } from './src/http_server/index.ts';
+// @ts-ignore
+import { wsInit, wss } from './src/ws/websocket.ts';
 import { mouse } from '@nut-tree/nut-js';
 
 const HTTP_PORT = 8181;
@@ -9,3 +10,10 @@ console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
 mouse.config.mouseSpeed = 200;
 wsInit();
+
+process.on('SIGINT', function () {
+  console.log('Closing http and ws connection');
+  wss.close();
+  httpServer.close();
+  process.exit();
+});
