+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.ts",
   "type": "module",
   "scripts": {
-    "start": "node ./index.js",
+    "start": "ts-node ./index.ts",
     "start:dev": "nodemon ./index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
